Update README.md
# DDD design parameters for ddd abstraction<br /><br />
Domain-Driven Design (DDD) is an approach to software development focusing on creating shared understanding of the domain problem between developers and domain experts. DDD provides a set of principles and practices to help design and build complex software systems to accurately reflect the real-world domain they are intended to model. 

key concepts in Domain-Driven Design:

Domain: The domain refers to the problem space or subject area that the software aims to address including relevant concepts, rules, and relationships defining the business or application domain.

Ubiquitous Language: DDD promotes the use of common, shared vocabulary, between developers and domain experts. This "ubiquitous language" ensures that both parties have a clear and consistent understanding of domain concepts, reducing misunderstandings and communication gaps.